Added feature baja de comentario
