+++ src/main/java/com/a2/backend/service/CommentService.java
@@ -2,6 +2,7 @@ package com.a2.backend.service;
 
 import com.a2.backend.entity.Comment;
 import com.a2.backend.model.CommentCreateDTO;
+
 import java.util.UUID;
 
 public interface CommentService {
@@ -11,4 +12,6 @@ public interface CommentService {
     Comment changeHighlight(UUID commentId);
 
     Comment changeHidden(UUID commentId);
+
+    Comment deleteComment(UUID id);
 }
