+++ src/test/java/com/a2/backend/controller/ProjectControllerActiveTest.java
@@ -1,10 +1,5 @@
 package com.a2.backend.controller;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import com.a2.backend.entity.Comment;
-import com.a2.backend.entity.Discussion;
 import com.a2.backend.entity.Project;
 import com.a2.backend.entity.User;
 import com.a2.backend.model.*;
@@ -12,10 +7,6 @@ import com.a2.backend.repository.DiscussionRepository;
 import com.a2.backend.repository.ProjectRepository;
 import com.a2.backend.repository.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-import java.util.UUID;
 import lombok.val;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,6 +15,15 @@ import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @AutoConfigureMockMvc
 public class ProjectControllerActiveTest extends AbstractControllerTest {
@@ -640,439 +640,5 @@ public class ProjectControllerActiveTest extends AbstractControllerTest {
                 .andExpect(status().isOk());
 
         assertEquals(4, user.getReputation());
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void
-            Test023_ProjectControllerWithValidCommentCreateDTOAndDiscussionIdWhenCreatingCommentThenStatusIsOk()
-                    throws Exception {
-        Discussion discussion =
-                projectRepository.findByTitle("TensorFlow").get().getDiscussions().get(0);
-
-        CommentCreateDTO commentCreateDTO =
-                CommentCreateDTO.builder().comment("test comment").build();
-
-        String contentAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                baseUrl + "/comment/" + discussion.getId())
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(objectMapper.writeValueAsString(commentCreateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        Comment comment = objectMapper.readValue(contentAsString, Comment.class);
-
-        assertEquals("test comment", comment.getComment());
-        assertEquals("rodrigo.pazos@ing.austral.edu.ar", comment.getUser().getEmail());
-        assertNotNull(comment.getDate());
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test024_ProjectControllerWithBlankCommentWhenCreatingCommentThenBadRequestIsReturned()
-            throws Exception {
-        Discussion discussion =
-                projectRepository.findByTitle("TensorFlow").get().getDiscussions().get(0);
-
-        CommentCreateDTO commentCreateDTO =
-                CommentCreateDTO.builder().comment("  \n  \r  \t  ").build();
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                baseUrl + "/comment/" + discussion.getId())
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(objectMapper.writeValueAsString(commentCreateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage).contains("Comment cannot be empty"));
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test025_ProjectControllerWithNullCommentWhenCreatingCommentThenBadRequestIsReturned()
-            throws Exception {
-        Discussion discussion =
-                projectRepository.findByTitle("TensorFlow").get().getDiscussions().get(0);
-
-        CommentCreateDTO commentCreateDTO = CommentCreateDTO.builder().comment(null).build();
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                baseUrl + "/comment/" + discussion.getId())
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(objectMapper.writeValueAsString(commentCreateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage).contains("Comment cannot be empty"));
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test026_ProjectControllerWithValidCommentIdWhenHighlightingShouldReturnHttpOkTest()
-            throws Exception {
-        val comment =
-                projectRepository
-                        .findByTitle("Kubernetes")
-                        .get()
-                        .getDiscussions()
-                        .get(0)
-                        .getComments()
-                        .get(0);
-
-        assertFalse(comment.isHighlighted());
-
-        String contentAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                baseUrl + "/highlight/" + comment.getId())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        val updatedComment = objectMapper.readValue(contentAsString, CommentDTO.class);
-
-        assertEquals(comment.getId(), updatedComment.getId());
-        assertTrue(updatedComment.isHighlighted());
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test027_ProjectControllerWithNotValidCommentIdWhenHighlightingShouldReturnBadRequest()
-            throws Exception {
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                baseUrl + "/highlight/" + UUID.randomUUID())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage).contains("Discussion with comment id"));
-    }
-
-    @Test
-    @WithMockUser(username = "agustin.ayerza@ing.austral.edu.ar")
-    void
-            Test028_ProjectControllerWithValidCommentIdButNotOwnerWhenHighlightingShouldReturnBadRequest()
-                    throws Exception {
-        val comment =
-                projectRepository
-                        .findByTitle("Kubernetes")
-                        .get()
-                        .getDiscussions()
-                        .get(0)
-                        .getComments()
-                        .get(0);
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                baseUrl + "/highlight/" + comment.getId())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage)
-                .contains("Only project owners can highlight comments"));
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test029_ProjectControllerWithValidCommentIdWhenHidingShouldReturnHttpOkTest()
-            throws Exception {
-        val comment =
-                projectRepository
-                        .findByTitle("Kubernetes")
-                        .get()
-                        .getDiscussions()
-                        .get(0)
-                        .getComments()
-                        .get(0);
-
-        assertFalse(comment.isHidden());
-
-        String contentAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(baseUrl + "/hide/" + comment.getId())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        val updatedComment = objectMapper.readValue(contentAsString, CommentDTO.class);
-
-        assertEquals(comment.getId(), updatedComment.getId());
-        assertTrue(updatedComment.isHidden());
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test030_ProjectControllerWithNotValidCommentIdWhenHidingShouldReturnBadRequest()
-            throws Exception {
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(baseUrl + "/hide/" + UUID.randomUUID())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage).contains("Discussion with comment id"));
-    }
-
-    @Test
-    @WithMockUser(username = "agustin.ayerza@ing.austral.edu.ar")
-    void Test031_ProjectControllerWithValidCommentIdButNotOwnerWhenHidingShouldReturnBadRequest()
-            throws Exception {
-        val comment =
-                projectRepository
-                        .findByTitle("Kubernetes")
-                        .get()
-                        .getDiscussions()
-                        .get(0)
-                        .getComments()
-                        .get(0);
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(baseUrl + "/hide/" + comment.getId())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage)
-                .contains("Only project owners can hide comments"));
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void
-            Test032_ProjectControllerWithNotValidDiscussionIdWhenGettingAllCommentsShouldReturnBadRequest()
-                    throws Exception {
-
-        String errorMessage =
-                mvc.perform(
-                                MockMvcRequestBuilders.get(
-                                                baseUrl + "/comments/" + UUID.randomUUID())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        assert (Objects.requireNonNull(errorMessage).contains("The discussion with id"));
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void
-            Test033_ProjectControllerWithValidDiscussionIdWhenGettingCommentsAsOwnerShouldReturnHttpOkTest()
-                    throws Exception {
-
-        val discussion = projectRepository.findByTitle("Kubernetes").get().getDiscussions().get(0);
-
-        String contentAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.get(
-                                                baseUrl + "/comments/" + discussion.getId())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        val comments = objectMapper.readValue(contentAsString, CommentDTO[].class);
-
-        assertNotNull(comments);
-        assertEquals(2, comments.length);
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void
-            Test034_ProjectControllerWithValidDiscussionIdWhenGettingCommentsAsCollaboratorShouldReturnHttpOkTest()
-                    throws Exception {
-
-        val discussion = projectRepository.findByTitle("Django").get().getDiscussions().get(0);
-
-        String contentAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.get(
-                                                baseUrl + "/comments/" + discussion.getId())
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        val comments = objectMapper.readValue(contentAsString, CommentDTO[].class);
-
-        assertNotNull(comments);
-        assertEquals(3, comments.length);
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void Test0035_ProjectControllerWhenReceivesValidUpdateDiscussionDTOshouldReturnOK()
-            throws Exception {
-
-        val project = projectRepository.findByTitle("Renovate");
-        String discussionTitle = "Discussion title";
-        List<String> discussionTags = Arrays.asList("desctag1", "desctag2");
-
-        DiscussionCreateDTO discussionCreateDTO =
-                DiscussionCreateDTO.builder()
-                        .forumTags(discussionTags)
-                        .title(discussionTitle)
-                        .build();
-
-        DiscussionCreateDTO updatedDiscussionCreateDTO =
-                DiscussionCreateDTO.builder()
-                        .forumTags(discussionTags)
-                        .title("AnotherName")
-                        .build();
-
-        String discussionAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.post(
-                                                "/project/"
-                                                        + project.get().getId().toString()
-                                                        + "/discussion")
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(
-                                                objectMapper.writeValueAsString(
-                                                        discussionCreateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isCreated())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-        DiscussionDTO createdDiscussion =
-                objectMapper.readValue(discussionAsString, DiscussionDTO.class);
-        DiscussionUpdateDTO discussionUpdateDTO =
-                DiscussionUpdateDTO.builder()
-                        .title("AnotherName")
-                        .forumTags(discussionTags)
-                        .build();
-
-        String updatedDiscussionAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                "/discussion/" + createdDiscussion.getId())
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(
-                                                objectMapper.writeValueAsString(
-                                                        discussionUpdateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-
-        String changedDiscussionAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.post(
-                                                "/project/"
-                                                        + project.get().getId().toString()
-                                                        + "/discussion")
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(
-                                                objectMapper.writeValueAsString(
-                                                        updatedDiscussionCreateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isBadRequest())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-    }
-
-    @Test
-    @WithMockUser(username = "rodrigo.pazos@ing.austral.edu.ar")
-    void
-            Test0036_ProjectControllerWhenReceivesValidUpdateDiscussionDTOButNotFromOwnerShouldReturnUnauthorized()
-                    throws Exception {
-        List<String> discussionTags = Arrays.asList("desctag1", "desctag2");
-        DiscussionUpdateDTO discussionUpdateDTO =
-                DiscussionUpdateDTO.builder()
-                        .title("AnotherName")
-                        .forumTags(discussionTags)
-                        .build();
-
-        String updatedDiscussionAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                "/discussion/"
-                                                        + discussionRepository
-                                                                .findByTitle(
-                                                                        "Issue with User ViewSet")
-                                                                .get()
-                                                                .getId())
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(
-                                                objectMapper.writeValueAsString(
-                                                        discussionUpdateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isUnauthorized())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
-    }
-
-    @Test
-    @WithMockUser(username = "agustin.ayerza@ing.austral.edu.ar")
-    void
-            Test0037_ProjectControllerWhenReceivesValidUpdateDiscussionDTOButNotFromOwnerShouldReturnUnauthorized()
-                    throws Exception {
-        List<String> discussionTags = Arrays.asList("desctag1", "desctag2");
-        DiscussionUpdateDTO discussionUpdateDTO =
-                DiscussionUpdateDTO.builder()
-                        .title("AnotherName")
-                        .forumTags(discussionTags)
-                        .build();
-        String updatedDiscussionAsString =
-                mvc.perform(
-                                MockMvcRequestBuilders.put(
-                                                "/discussion/"
-                                                        + discussionRepository
-                                                                .findByTitle(
-                                                                        "Issue with User ViewSet")
-                                                                .get()
-                                                                .getId())
-                                        .contentType(MediaType.APPLICATION_JSON)
-                                        .content(
-                                                objectMapper.writeValueAsString(
-                                                        discussionUpdateDTO))
-                                        .accept(MediaType.APPLICATION_JSON))
-                        .andExpect(status().isOk())
-                        .andReturn()
-                        .getResponse()
-                        .getContentAsString();
     }
 }
